Layout, Pagination 컴포넌트와 themeContext에 typesript 적용
