+++ src/components/Pagination/index.tsx
@@ -14,7 +14,12 @@ const StyledLink = styled(Link)`
   font-weight: ${({ fontWeight }) => fontWeight};
 `;
 
-const Pagination = ({ totalPages, currentPage }) => {
+type Props = {
+  totalPages: number;
+  currentPage: number;
+};
+
+const Pagination = ({ totalPages, currentPage }: Props) => {
   const items = Array.from(Array(totalPages).keys());
 
   return (
