--- src/contexts/themeContext.js
@@ -1,27 +1,0 @@
-import React, { createContext, useState } from 'react';
-
-const ThemeContext = createContext({
-  theme: 'light',
-  toggleTheme: null
-});
-
-const ThemeProvider = ({ children }) => {
-  const [theme, setTheme] = useState('dark');
-
-  const toggleTheme = () => {
-    setTheme(theme === 'dark' ? 'light' : 'dark');
-  };
-
-  return (
-    <ThemeContext.Provider
-      value={{
-        theme,
-        toggleTheme
-      }}
-    >
-      {children}
-    </ThemeContext.Provider>
-  );
-};
-
-export { ThemeContext, ThemeProvider };
